requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Trading and Finance
python-binance>=1.0.19
ccxt>=4.0.0
ta-lib>=0.4.28
yfinance>=0.2.18

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.1.0
ray[rllib]>=2.7.0

# Optimization
optuna>=3.3.0
hyperopt>=0.2.7

# Natural Language Processing
transformers>=4.30.0
sentence-transformers>=2.2.2
nltk>=3.8.1
textblob>=0.17.1

# API and HTTP
requests>=2.31.0
aiohttp>=3.8.5
websockets>=11.0.3

# Data Processing
polars>=0.19.0
dask>=2023.8.0
vaex>=4.17.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# Logging and Monitoring
wandb>=0.15.0
tensorboard>=2.13.0
prometheus-client>=0.17.0

# Explainability
shap>=0.42.0
lime>=0.2.0.1
interpret>=0.4.0

# Configuration and Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.4.0

# Database
sqlalchemy>=2.0.0
redis>=4.6.0
influxdb-client>=1.36.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# System and Performance
psutil>=5.9.0
memory-profiler>=0.61.0
line-profiler>=4.1.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.1

# Async and Concurrency
asyncio-mqtt>=0.13.0
aioredis>=2.0.0

# Time and Date
python-dateutil>=2.8.2
pytz>=2023.3

# Statistics and Analysis
statsmodels>=0.14.0
arch>=6.2.0
empyrical>=0.5.5

# Optional: GPU support (uncomment if needed)
# cupy-cuda11x>=12.0.0  # For CUDA 11.x
# cupy-cuda12x>=12.0.0  # For CUDA 12.x

# Optional: Advanced ML
# xgboost>=1.7.0
# lightgbm>=4.0.0
# catboost>=1.2.0

# Optional: Web Framework (for dashboard)
# fastapi>=0.100.0
# uvicorn>=0.23.0
# streamlit>=1.25.0

# Optional: Message Queues
# pika>=1.3.0
# kafka-python>=2.0.2

# Optional: Cloud Services
# boto3>=1.28.0  # AWS
# google-cloud-storage>=2.10.0  # GCP
# azure-storage-blob>=12.17.0  # Azure 